fix abs paths for json generator
